# Improve RubyHash#resize performance